+++ WebClient/FileManager.ts
@@ -30,6 +30,7 @@ import {
   second,
   ResolvedConflict,
   DEFAULT_USER_MODE,
+  DEFAULT_SCENE_BACKGROUND_COLOR,
 } from 'common/Files';
 import { 
   filterFilesBySelection, 
@@ -186,6 +187,14 @@ export class FileManager {
       return;
     }
     var objs = this.objects.filter(o => o.id === this._appManager.user.username);
+    if (objs.length > 0) {
+      return objs[0];
+    }
+    return null;
+  }
+
+  get globalsFile(): Object {
+    let objs = this.objects.filter((o => o.id === 'globals'));
     if (objs.length > 0) {
       return objs[0];
     }
@@ -497,6 +506,7 @@ export class FileManager {
 
     this._setupOffline();
     await this._initUserFile();
+    await this._initGlobalsFile();
 
     // Replay the existing files for the components that need it this way
     const filesState = this._files.store.state();
@@ -545,6 +555,19 @@ export class FileManager {
         _position: { x: 0, y: 0, z: 0},
         _mode: DEFAULT_USER_MODE,
         _workspace: null
+      });
+    }
+  }
+
+  private async _initGlobalsFile() {
+    this._setStatus('Updating globals file...');
+    let globalsFile = this.globalsFile;
+    if (!globalsFile) {
+      await this.createFile('globals', {
+        _hidden: true,
+        _workspace: null,
+        _position: { x:0, y: 0, z: 0},
+        _sceneBackgroundColor: DEFAULT_SCENE_BACKGROUND_COLOR
       });
     }
   }