+++ common/Files/File.ts
@@ -20,7 +20,7 @@ export interface Object {
         _movable?: boolean;
         _editingFile?: string;
         ['stroke.color']?: string;
-        ['stroke.linewidth']?: number;
+        ['stroke.width']?: number;
         ['line.to']?: string;
         ['line.color']?: string;
         ['label']?: string;
@@ -52,6 +52,7 @@ export interface Workspace {
     scale: number | null;
     gridScale: number | null;
     defaultHeight: number | null;
+    color: string;
 }
 
 /**
@@ -86,7 +87,7 @@ export interface PartialFile {
         _movable?: boolean;
         _editingFile?: string;
         ['stroke.color']?: string;
-        ['stroke.linewidth']?: number;
+        ['stroke.width']?: number;
         ['line.to']?: string;
         ['line.color']?: string;
         ['label']?: string;
@@ -103,6 +104,7 @@ export interface PartialFile {
     },
     scale?: number;
     gridScale?: number;
+    color?: string;
 }
 
 /**
@@ -139,3 +141,13 @@ export const DEFAULT_WORKSPACE_HEIGHT_INCREMENT = 0.1;
  * The minimum height that hexes in a workspace can be.
  */
 export const DEFAULT_WORKSPACE_MIN_HEIGHT = 0.1;
+
+/**
+ * The default color for workspaces.
+ */
+export const DEFAULT_WORKSPACE_COLOR = "#999999";
+
+/**
+ * The default color for scene background.
+ */
+export const DEFAULT_SCENE_BACKGROUND_COLOR = "#CCE6FF";